feat(pvm): support the 64-bit isa (#50)
* feat(pvm): introduce isa for gp@0.5.4

* feat(pvm): use 64-bits for isa

* feat(testing): update specjam to 0.0.7

* feat(pvm): introduce 64-bit arthmetic visitors

* feat(pvm): introduce store ind related instructions

* feat(pvm/parser): introduce custom result

* feat(pvm): introduce memory guard

* feat(pvm): use read interface for reading from memory

* feat(pvm): introduce memory write interface

* feat(pvm): introduce memory write interface

* feat(pvm): remove multiple pages access and hack u64 encoding

* chore(pvm): clean instruction reader

* feat(pvm): support jump table

* feat(pvm): support page address on page fault

* feat(pvm): minor docs

+++ crates/pvm/parser/src/format/ii.rs
@@ -1,3 +1,5 @@
+//! The II instruction format.
+
 use crate::format::{II, ISA};
 
 impl From<II> for Vec<u8> {
@@ -28,8 +30,8 @@ impl From<&[u8]> for II {
         let mid = 1 + x_len as usize;
 
         II {
-            imm0: u32::read_imm(&bytes[1..mid]),
-            imm1: u32::read_imm(&bytes[mid..]),
+            imm0: u64::read_imm(&bytes[1..mid]),
+            imm1: u64::read_imm(&bytes[mid..]),
         }
     }
 }
